refactor(console): give the fetch validators their own modules under the health gate
commands-validate.ts returns to command validation only (now 10.0,
from its 9.09 baseline: typed NameCheck, one-shape cooldown range,
data-driven trailer/control checks); the fetch-def validators move to
fetch-validate.ts and the token/slug grammar to fetch-tokens.ts, both
scoring 10.0, with re-export chains keeping every existing import path
working.
